AI/ml_pipeline/requirements.txt
# Core Dependencies
langgraph>=0.0.30
langchain>=0.1.0
langchain-aws>=0.1.0

# MLflow for Experiment Tracking
mlflow>=2.10.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Data Science Stack
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4.0
scipy>=1.11.0

# Preprocessing Techniques (for 28 technique implementations)
category-encoders>=2.6.0  # For target encoding, binary encoding, hash encoding
imbalanced-learn>=0.12.0  # For SMOTE and other imbalanced data handling
statsmodels>=0.14.0  # For advanced statistical imputation (MICE)

# Natural Language Processing & Embeddings
sentence-transformers>=2.2.2

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Configuration Management
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Logging and Monitoring
loguru>=0.7.0

# Type Checking
typing-extensions>=4.9.0

# Data Validation
pydantic>=2.5.0

# HTTP Requests
requests>=2.31.0

# Web Framework (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Progress Bars
tqdm>=4.66.0

# Optional: Advanced Algorithms
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0

# Optional: Deep Learning
# torch>=2.1.0
# tensorflow>=2.15.0

# Development Dependencies (install with pip install -r requirements-dev.txt)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# pytest-mock>=3.12.0
# black>=23.0.0
# flake8>=6.1.0
# mypy>=1.7.0
# isort>=5.13.0
# pre-commit>=3.6.0
